// apps/nfid-frontend/src/integration/nft/impl/icpswap/idl/SwapNFT.d.ts
export type AccountIdentifier = string;
export type AccountIdentifier__1 = string;
export interface AllowanceRequest {
  token: TokenIdentifier;
  owner: User;
  spender: Principal;
}
export interface ApproveAllRequest {
  approved: boolean;
  spender: User;
}
export interface ApproveRequest {
  token: TokenIdentifier;
  subaccount: [] | [SubAccount];
  allowance: Balance;
  spender: Principal;
}
export type Balance = bigint;
// ...
export type Memo = Array<number>;
// ...
export type Remark = string;
// ...
export type TokenIndex__1 = number;
export type TransType = { burn: null } | { mint: null } | { approve: null } | { transfer: null };
export interface TransferRecord {
  to: AccountIdentifier;
  remark: Remark;
  tokenId: TokenIndex__1;
  from: AccountIdentifier;
  hash: string;
  memo: [] | [Memo];
  time: bigint;
  tokenName: string;
  txType: TransType;
  caller: AccountIdentifier;
  price: Balance;
  amount: Balance;
}